+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Customizable full player action buttons via "more"
+- Scroll thumb customization
 
 ### Fixed